+++ activerecord/test/cases/query_cache_test.rb
@@ -141,6 +141,79 @@ class QueryCacheTest < ActiveRecord::TestCase
     clean_up_connection_handler
   end
 
+  def test_cache_is_not_applied_when_config_is_false
+    ActiveRecord::Base.connected_to(role: :reading) do
+      db_config = ActiveRecord::Base.configurations.configs_for(env_name: "arunit", name: "primary")
+      ActiveRecord::Base.establish_connection(db_config.configuration_hash.merge(query_cache: false))
+    end
+
+    mw = middleware { |env|
+      ActiveRecord::Base.connection_handler.connection_pool_list(:reading).each do |pool|
+        assert_not_predicate pool.lease_connection, :query_cache_enabled
+        assert_nil pool.query_cache.instance_variable_get(:@max_size)
+      end
+    }
+
+    mw.call({})
+  ensure
+    clean_up_connection_handler
+  end
+
+  def test_cache_is_applied_when_config_is_string
+    ActiveRecord::Base.connected_to(role: :reading) do
+      db_config = ActiveRecord::Base.configurations.configs_for(env_name: "arunit", name: "primary")
+      ActiveRecord::Base.establish_connection(db_config.configuration_hash.merge(query_cache: "unlimited"))
+    end
+
+    mw = middleware { |env|
+      ActiveRecord::Base.connection_handler.connection_pool_list(:reading).each do |pool|
+        assert_predicate pool.lease_connection, :query_cache_enabled
+        assert_nil pool.query_cache.instance_variable_get(:@max_size)
+      end
+    }
+
+    mw.call({})
+  ensure
+    clean_up_connection_handler
+  end
+
+  def test_cache_is_applied_when_config_is_integer
+    ActiveRecord::Base.connected_to(role: :reading) do
+      db_config = ActiveRecord::Base.configurations.configs_for(env_name: "arunit", name: "primary")
+      ActiveRecord::Base.establish_connection(db_config.configuration_hash.merge(query_cache: 42))
+    end
+
+    mw = middleware { |env|
+      ActiveRecord::Base.connection_handler.connection_pool_list(:reading).each do |pool|
+        assert_predicate pool.lease_connection, :query_cache_enabled
+        assert_equal 42, pool.query_cache.instance_variable_get(:@max_size)
+      end
+    }
+
+    mw.call({})
+  ensure
+    clean_up_connection_handler
+  end
+
+  def test_cache_is_applied_when_config_is_nil
+    ActiveRecord::Base.connected_to(role: :reading) do
+      db_config = ActiveRecord::Base.configurations.configs_for(env_name: "arunit", name: "primary")
+      ActiveRecord::Base.establish_connection(db_config.configuration_hash.merge(query_cache: nil))
+    end
+
+    mw = middleware { |env|
+      ActiveRecord::Base.connection_handler.connection_pool_list(:reading).each do |pool|
+        assert_predicate pool.lease_connection, :query_cache_enabled
+      end
+    }
+
+    mw.call({})
+  ensure
+    clean_up_connection_handler
+  end
+
+
+
   if Process.respond_to?(:fork) && !in_memory_db?
     def test_query_cache_with_forked_processes
       ActiveRecord::Base.connected_to(role: :reading) do